requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.3
djangorestframework==3.12.2
gunicorn==20.0.4
markdown2==2.3.9
peewee==3.13.3
protobuf==3.12.4
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.2.0
pip==20.2.4
win32-setctime==1.0.3
app==0.0.1
